Convert worker service to web API
