+++ IYSIntegration.Scheduled/Models/BatchItemResult.cs
@@ -1,4 +1,4 @@
-﻿namespace IYSIntegration.WorkerService.Models
+﻿namespace IYSIntegration.Scheduled.Models
 {
     public class BatchItemResult
     {
