+++ IYSIntegration.Scheduled/Models/SfConsentResult.cs
@@ -1,4 +1,4 @@
-﻿namespace IYSIntegration.WorkerService.Models
+﻿namespace IYSIntegration.Scheduled.Models
 {
     public class SfConsentResult
     {
