+++ IYSIntegration.Scheduled/Properties/PublishProfiles/FolderProfile.pubxml
