+++ IYSIntegration.Scheduled/Utilities/IntegrationHelper.cs
@@ -9,7 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace IYSIntegration.WorkerService.Utilities
+namespace IYSIntegration.Scheduled.Utilities
 {
     public class IntegrationHelper : IIntegrationHelper
     {
