+++ IYSIntegration.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34009.444
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IYSIntegration.WorkerService", "IYSIntegration.WorkerService\IYSIntegration.WorkerService.csproj", "{0E9781E2-9B03-4D58-9385-3C2B159BBD02}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IYSIntegration.Scheduled", "IYSIntegration.Scheduled\IYSIntegration.Scheduled.csproj", "{0E9781E2-9B03-4D58-9385-3C2B159BBD02}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IYSIntegration.Common", "IYSIntegration.Common\IYSIntegration.Common.csproj", "{FF28D53A-BF37-4B27-9902-30389AF9AD3A}"
 EndProject
