+++ README.md
@@ -1,1 +1,11 @@
 # IYSRestruct
+
+## Delay Settings
+
+The following values from `appsettings.json` indicate how many minutes each delay represents:
+
+- `MultipleConsentQueryDelay` – 30 seconds (0.5 minutes)
+- `MultipleConsentRequestDelay` – 60 seconds (1 minute)
+- `SingleConsentQueryDelay` – 300 seconds (5 minutes)
+- `PullConsentQueryDelay` – 3600 seconds (60 minutes)
+- `SfAddConsentDelay` – 90 seconds (1.5 minutes)
